requirements.txt
fastapi
uvicorn[standard]
openai
httpx
websockets
python-dotenv
plivo
asyncpg
redis[hiredis]
